+++ db/migrate/20150214151020_add_city_and_gender_to_users.rb
@@ -1,0 +1,12 @@
+class AddCityAndGenderToUsers < ActiveRecord::Migration
+  def self.up
+    add_column :users, :city, :string
+    add_column :users, :gender, :integer, limit: 1
+    add_index :users, :city
+  end
+  def self.down
+    remove_index :users, :city
+    remove_column :users, :city
+    remove_column :users, :gender
+  end
+end