+++ db/migrate/20150214152225_add_socials_to_users.rb
@@ -1,0 +1,8 @@
+class AddSocialsToUsers < ActiveRecord::Migration
+  def self.up
+    add_column :users, :socials, :hstore
+  end
+  def self.down
+    remove_column :users, :socials
+  end
+end